Refactor scan_hosts to works with any esp_netif interface
