+++ src/modules/wifi/deauther.h
@@ -1,0 +1,8 @@
+#ifndef WIFI_DEAUTHER_H
+#define WIFI_DEAUTHER_H
+
+#include "scan_hosts.h"
+
+void stationDeauth(Host host);
+
+#endif
